Added most of create_room command
It's a bit hard to test without the join chat room command, but most of
the functionality for the create room command should be added.  It
currently does not add the user to the room, though there is theoretical
code for this.  Will we manually edit data to add a user to a room, or
will we have a function?