Added whisper and reply functionality
Filled in the whisper and relpy function stubs and tested them.  For
both whisper and reply, the message can contain no spaces, but otherwise
they work perfectly.  I chose to put parenthesis around the sent message
to indicate it was a whisper, but we can talk about this next class.